Completed most of Register Use-case, set patterns, changed .mdj and servlet mappings
only validateUserDetails remaining.
This pattern is to be followed for other InterfaceClasses as well: the
display feature will be in the Interface Class, and the .jsp files will
only redirect to it. This allows for greater flexibility for other
inputs.
The static pages like about-us, contact-us and home, can remain .jsp.

Also included in this commit are several servlet mappings: All
IC-classes will have the URL:    /<whatever>Handler   The .jsp files
which call these servlets will have mappings for   /<whatever>   and
/<whatever>.jsp